Moved list of macros to the new component
