+++ src/components/macro/edit/macro-edit.component.scss
@@ -1,0 +1,5 @@
+.not-found {
+    margin-top: 30px;
+    font-size: 16px;
+    text-align: center;
+}
