+++ src/components/macro/list/macro-list.component.scss
@@ -1,4 +1,4 @@
-@import '../../main-app/global-styles';
+@import '../../../main-app/global-styles';
 
 :host {
     display: flex;
@@ -155,10 +155,4 @@ h1 {
         -ms-user-select: none;
         user-select: none;
     }
-}
-
-.not-found {
-    margin-top: 30px;
-    font-size: 16px;
-    text-align: center;
 }
